requirements.txt
marshmallow
pytest>=7.0.0
pytest-subtests
pyyaml
pyserial
filelock
